refactor(metal-vm): Split the collaborator contracts out of the runtime file
`runtime.go` held the firecracker contract next to the network and storage ones. Each collaborator now has its own file, so the runtime file names only what firecracker implements.

- Move `Network`, `NetworkRequest`, `NetworkReleaseRequest`, and `NetworkInterface` to `network.go`.
- Move `Storage`, `DiskUsage`, `Snapshots`, `SnapshotRequest`, and `StagedSnapshot` to `storage.go`.
