+++ metal/internal/vm/network.go
@@ -1,0 +1,33 @@
+package vm
+
+import "context"
+
+// NetworkInterface contains the host network values used by a runtime.
+type NetworkInterface struct {
+	NetworkNamespacePath string
+	TapName              string
+	MACAddress           string
+	GuestIPAddress       string
+	GatewayIPAddress     string
+}
+
+// Network converges and releases host network resources.
+type Network interface {
+	Ensure(context.Context, NetworkRequest) (NetworkInterface, error)
+	Release(context.Context, NetworkReleaseRequest) error
+}
+
+// NetworkRequest contains the complete desired host network state.
+type NetworkRequest struct {
+	VirtualMachineID string
+	UserID           uint32
+	GroupID          uint32
+	Configuration    NetworkConfiguration
+}
+
+// NetworkReleaseRequest identifies host network resources to remove.
+type NetworkReleaseRequest struct {
+	VirtualMachineID  string
+	UserID            uint32
+	WireGuardMeshIPv6 string
+}
